+++ frontend/src/scenes/frame/frameLogic.ts
@@ -13,105 +13,30 @@ import { entityImagesModel } from '../../models/entityImagesModel'
 import { arrangeNodes } from '../../utils/arrangeNodes'
 import { isInFrameAdminMode } from '../../utils/frameAdmin'
 import { secureToken } from '../../utils/secureToken'
-import versions from '../../../../versions.json'
+import {
+  type ChangeDetail,
+  CURRENT_FRAMEOS_VERSION,
+  type DeployPlanResponse,
+  type DeployRecommendation,
+  type SummaryItem,
+  buildDeployPlanRequestBody,
+  buildDeployRecommendation,
+  buildFastDeployPlanSummary,
+  buildFullDeployPlanSummary,
+} from './frameDeployUtils'
+import { getDeployPlanErrorMessage } from './frameDeployErrors'
 
-export interface FrameLogicProps {
-  frameId: number
-}
-
-export interface ChangeDetail {
-  label: string
-  requiresFullDeploy: boolean
-}
-
-export interface SummaryItem {
-  label: string
-  value: string
-}
-
-export interface FastDeployPlanResponse {
-  reload_supported: boolean
-  tls_settings_changed: boolean
-  action: string
-}
-
-export interface FullDeployPlanResponse {
-  target: {
-    arch: string
-    distro: string
-    version: string
-    total_memory_mb: number
-  }
-  low_memory: boolean
-  drivers: string[]
-  binary: {
-    will_attempt_cross_compile?: boolean
-    cross_compile_supported?: boolean
-    build_host_configured?: boolean
-    prebuilt_target?: string | null
-    has_prebuilt_entry?: boolean
-  }
-  packages: {
-    name: string
-    reason: string
-    installed: boolean
-    needs_install: boolean
-  }[]
-  package_alternatives: {
-    names: string[]
-    reason: string
-    installed_package?: string | null
-    satisfied: boolean
-  }[]
-  lgpio: {
-    required: boolean
-    installed: boolean
-  }
-  quickjs: {
-    required_if_remote_build: boolean
-    dirname?: string | null
-    installed: boolean
-  }
-  ssh_keys_need_install: boolean
-  post_deploy?: {
-    i2c?: {
-      needs_boot_config_line?: boolean
-      needs_runtime_enable?: boolean
-    }
-    spi_action?: 'enable' | 'disable' | 'unchanged'
-    reboot_schedule?: {
-      needs_update?: boolean
-      needs_remove?: boolean
-    }
-    bootconfig_changes?: { action: 'add' | 'remove'; line: string }[]
-    disable_userconfig?: boolean
-    final_action?: 'reboot' | 'restart_frameos'
-  }
-}
-
-export interface DeployPlanResponse {
-  mode: 'fast' | 'full' | 'combined'
-  frame_id: number
-  frame_name: string
-  build_id: string
-  previous_frameos_version?: string | null
-  notes: string[]
-  fast_deploy?: FastDeployPlanResponse | null
-  full_deploy?: FullDeployPlanResponse | null
-}
+export type { ChangeDetail, DeployPlanResponse, DeployRecommendation, SummaryItem } from './frameDeployUtils'
 
 interface DeployPlanApiResponse {
   plan: DeployPlanResponse
 }
 
-export interface DeployRecommendation {
-  mode: 'fast' | 'full'
-  title: string
-  description: string
+export interface FrameLogicProps {
+  frameId: number
 }
 
 const DEFAULT_BROWSER_TITLE = 'FrameOS Backend'
-const CURRENT_FRAMEOS_VERSION = (versions.frameos || 'dev').split('+')[0]
 
 function setBrowserTitle(frame?: FrameType | null): void {
   if (typeof document === 'undefined') {
@@ -469,197 +394,6 @@ function buildUndeployedSummaryItems(
   return items
 }
 
-function buildDeployPlanRequestBody(frame: Partial<FrameType>): Record<string, any> {
-  const json: Record<string, any> = {}
-  for (const key of FRAME_KEYS) {
-    json[key] = frame[key]
-  }
-  return json
-}
-
-function getResponseDetail(payload: unknown): string | null {
-  if (payload && typeof payload === 'object' && 'detail' in payload && typeof payload.detail === 'string') {
-    return payload.detail
-  }
-  return null
-}
-
-function isFrameConnectionError(detail: string | null | undefined): boolean {
-  if (!detail) {
-    return false
-  }
-
-  const normalizedDetail = detail.toLowerCase()
-  return (
-    normalizedDetail.includes('unable to connect to') ||
-    normalizedDetail.includes('unable to reach frame') ||
-    normalizedDetail.includes('agent disconnected') ||
-    normalizedDetail.includes('connection refused') ||
-    normalizedDetail.includes('no route to host') ||
-    normalizedDetail.includes('name or service not known') ||
-    normalizedDetail.includes('temporary failure in name resolution') ||
-    normalizedDetail.includes('timeout to http') ||
-    normalizedDetail.includes('timed out')
-  )
-}
-
-function buildFrameosVersionSummary(plan?: DeployPlanResponse | null): SummaryItem[] {
-  if (!plan) {
-    return []
-  }
-
-  const previousVersion = plan.previous_frameos_version ? String(plan.previous_frameos_version).split('+')[0] : 'Not deployed'
-  return [
-    {
-      label: 'FrameOS version',
-      value: previousVersion === CURRENT_FRAMEOS_VERSION ? CURRENT_FRAMEOS_VERSION : `${previousVersion} -> ${CURRENT_FRAMEOS_VERSION}`,
-    },
-  ]
-}
-
-function buildFastDeployPlanSummary(plan?: DeployPlanResponse | null): SummaryItem[] {
-  const fastPlan = plan?.fast_deploy
-  if (!fastPlan) {
-    return []
-  }
-
-  if (!fastPlan.tls_settings_changed) {
-    return []
-  }
-
-  return [{ label: 'Fast deploy behavior', value: 'Restart FrameOS because TLS settings changed' }]
-}
-
-function buildFullDeployPlanSummary(plan?: DeployPlanResponse | null, frame?: Partial<FrameType> | null): SummaryItem[] {
-  const fullPlan = plan?.full_deploy
-  if (!fullPlan) {
-    return []
-  }
-
-  const packagesToInstall = fullPlan.packages.filter((pkg) => pkg.needs_install).map((pkg) => pkg.name)
-  const previousVersion = plan.previous_frameos_version ? String(plan.previous_frameos_version).split('+')[0] : 'Not deployed'
-  const items: SummaryItem[] = [
-    {
-      label: 'FrameOS version',
-      value: previousVersion === CURRENT_FRAMEOS_VERSION ? CURRENT_FRAMEOS_VERSION : `${previousVersion} -> ${CURRENT_FRAMEOS_VERSION}`,
-    },
-    ...(frame?.device ? [{ label: 'Device', value: String(frame.device) }] : []),
-    {
-      label: 'Target',
-      value: `${fullPlan.target.distro} ${fullPlan.target.version} · ${fullPlan.target.arch} · ${fullPlan.target.total_memory_mb} MiB`,
-    },
-    {
-      label: 'Build strategy',
-      value: fullPlan.binary.will_attempt_cross_compile
-        ? fullPlan.binary.build_host_configured
-          ? 'Cross-compile on the configured build host'
-          : 'Cross-compile locally on this server'
-        : fullPlan.binary.cross_compile_supported
-        ? 'Build on device because cross-compilation is disabled'
-        : 'Build on device because cross-compilation is unavailable for this target',
-    },
-  ]
-
-  if (fullPlan.drivers.length > 0) {
-    items.push({ label: 'Drivers', value: stringifyList(fullPlan.drivers) })
-  }
-  if (packagesToInstall.length > 0) {
-    items.push({ label: 'Packages to install', value: stringifyList(packagesToInstall) })
-  }
-  if (!fullPlan.binary.will_attempt_cross_compile && fullPlan.quickjs.required_if_remote_build && !fullPlan.quickjs.installed) {
-    items.push({
-      label: 'QuickJS',
-      value: `${fullPlan.quickjs.dirname || 'Required'} will be prepared for the on-device build`,
-    })
-  }
-  if (fullPlan.lgpio.required && !fullPlan.lgpio.installed) {
-    items.push({ label: 'lgpio', value: 'Will be installed for the selected drivers' })
-  }
-  if (fullPlan.ssh_keys_need_install) {
-    items.push({ label: 'SSH keys', value: 'Selected deploy keys will be installed on the frame' })
-  }
-  if (fullPlan.post_deploy?.bootconfig_changes?.length) {
-    items.push({
-      label: 'Boot config',
-      value: fullPlan.post_deploy.bootconfig_changes
-        .map((change) => `${change.action === 'add' ? 'Add' : 'Remove'} ${change.line}`)
-        .join(', '),
-    })
-  }
-  if (fullPlan.post_deploy?.i2c?.needs_boot_config_line || fullPlan.post_deploy?.i2c?.needs_runtime_enable) {
-    items.push({ label: 'I2C', value: 'Will be enabled for the selected drivers' })
-  }
-  if (fullPlan.post_deploy?.spi_action === 'enable') {
-    items.push({ label: 'SPI', value: 'Will be enabled for the selected drivers' })
-  }
-  if (fullPlan.post_deploy?.spi_action === 'disable') {
-    items.push({ label: 'SPI', value: 'Will be disabled for the selected drivers' })
-  }
-  if (fullPlan.post_deploy?.disable_userconfig) {
-    items.push({ label: 'User config', value: 'First-deploy setup will disable the system userconfig service' })
-  }
-  if (fullPlan.post_deploy?.reboot_schedule?.needs_update) {
-    items.push({ label: 'Reboot schedule', value: 'Scheduled reboot config will be installed or updated' })
-  }
-  if (fullPlan.post_deploy?.reboot_schedule?.needs_remove) {
-    items.push({ label: 'Reboot schedule', value: 'Old scheduled reboot config will be removed' })
-  }
-  if (fullPlan.low_memory && !fullPlan.binary.will_attempt_cross_compile) {
-    items.push({ label: 'Low memory', value: 'FrameOS will be stopped before the on-device build' })
-  }
-  if (fullPlan.post_deploy?.final_action === 'reboot') {
-    items.push({ label: 'After deploy', value: 'Device reboot required' })
-  }
-
-  return items
-}
-
-function buildDeployRecommendation(
-  plan: DeployPlanResponse | null,
-  hasPreviousDeploy: boolean,
-  deployChangeDetails: ChangeDetail[]
-): DeployRecommendation | null {
-  if (!plan) {
-    return null
-  }
-
-  const previousVersion = plan.previous_frameos_version ? String(plan.previous_frameos_version).split('+')[0] : null
-  const versionChanged = previousVersion !== CURRENT_FRAMEOS_VERSION
-  const fullDeployChanges = deployChangeDetails
-    .filter((change) => change.requiresFullDeploy && !change.label.startsWith('FrameOS upgrade'))
-    .map((change) => change.label)
-
-  if (!hasPreviousDeploy) {
-    return {
-      mode: 'full',
-      title: 'Suggested: full deploy',
-      description: 'This frame has not been deployed yet, so FrameOS, dependencies, and system changes need a full deploy.',
-    }
-  }
-
-  if (fullDeployChanges.length > 0) {
-    return {
-      mode: 'full',
-      title: 'Suggested: full deploy',
-      description: `These changes require rebuilding or reinstalling FrameOS: ${fullDeployChanges.join(', ')}.`,
-    }
-  }
-
-  if (versionChanged) {
-    return {
-      mode: 'fast',
-      title: 'Suggested: fast deploy',
-      description: `Fast deploy is enough to push the latest frame config and interpreted scenes. Use full deploy only if you also want to update the FrameOS runtime from ${previousVersion ?? 'unknown'} to ${CURRENT_FRAMEOS_VERSION}.`,
-    }
-  }
-
-  return {
-    mode: 'fast',
-    title: 'Suggested: fast deploy',
-    description: 'No pending changes require rebuilding FrameOS, so a fast deploy (reload) is enough to bring the frame up to date.',
-  }
-}
-
 async function resolveTemplateImageUrl(template: Partial<TemplateType>): Promise<string | null> {
   if (template.id) {
     return `/api/templates/${template.id}/image`
@@ -913,7 +647,7 @@ export const frameLogic = kea<frameLogicType>([
         })),
       }),
       submit: async (frame) => {
-        const json = buildDeployPlanRequestBody(frame)
+        const json = buildDeployPlanRequestBody(frame, FRAME_KEYS)
         if (values.nextAction) {
           json['next_action'] = values.nextAction
         }
@@ -1076,14 +810,11 @@ export const frameLogic = kea<frameLogicType>([
       const response = await apiFetch(`/api/frames/${values.frameId}/deploy_plan`, {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(buildDeployPlanRequestBody(values.frameForm)),
+        body: JSON.stringify(buildDeployPlanRequestBody(values.frameForm, FRAME_KEYS)),
       })
       if (!response.ok) {
         const payload = await response.json().catch(() => ({}))
-        const detail = getResponseDetail(payload)
-        actions.loadDeployPlansFailure(
-          isFrameConnectionError(detail) ? 'Failed to conennect to frame' : detail || 'Failed to load deploy plans'
-        )
+        actions.loadDeployPlansFailure(getDeployPlanErrorMessage(payload))
         return
       }
